requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.6.0
fonttools==4.29.1
gunicorn @ git+https://github.com/benoitc/gunicorn.git@1aae54a8e1a1b3022fbce42259b3bf36e2808f07
httpie==3.0.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.1
open3d==0.14.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pygments==2.11.2
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.2
trimesh==3.10.0
urllib3==1.26.8
virtualenv==20.13.2
virtualenv-clone==0.5.7
whitenoise==6.0.0
